Proyecto TCC/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
openpyxl>=3.1.0
streamlit-authenticator>=0.4.2
pyyaml>=6.0.0
requests>=2.31.0 